requirements.txt
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.35.0
django-bootstrap-toolkit==2.15.0
django-bootstrap4==0.0.6
django-contrib-comments==1.8.0
django-crispy-forms==1.7.1
django-filter==1.1.0
django-fluent-comments==2.0
django-registration-redux==1.4
django-tag-parser==3.1
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
pycparser==2.18
PyJWT==1.6.0
python-akismet==0.3.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
wincertstore==0.2
